refactor(core): extract shared kernel module
Move frontmatter, validated types, and DeviceContext into shared/
module to establish a shared kernel boundary for cross-cutting
domain types.
