+++ core/src/format.rs
@@ -1,23 +1,8 @@
 use chrono::{DateTime, FixedOffset, Local};
-use serde::Serialize;
 
 use crate::error::CoreError;
-use crate::frontmatter::{self, ContextMeta, NoteFrontmatter};
-
-#[derive(Debug, Clone, Serialize)]
-pub struct DeviceContext {
-    pub battery: u8,
-    pub is_charging: bool,
-}
-
-impl DeviceContext {
-    pub fn mock() -> Self {
-        Self {
-            battery: 50,
-            is_charging: false,
-        }
-    }
-}
+use crate::shared::context::DeviceContext;
+use crate::shared::frontmatter::{self, ContextMeta, NoteFrontmatter};
 
 pub fn format_timeline_line(
     text: &str,
@@ -53,7 +38,7 @@ pub fn format_note_markdown(
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::frontmatter::NoteFrontmatter;
+    use crate::shared::frontmatter::NoteFrontmatter;
     use chrono::TimeZone;
 
     fn fixed_timestamp() -> DateTime<Local> {
@@ -65,13 +50,6 @@ mod tests {
             battery: 82,
             is_charging: false,
         }
-    }
-
-    #[test]
-    fn test_device_context_mock() {
-        let ctx = DeviceContext::mock();
-        assert_eq!(ctx.battery, 50);
-        assert!(!ctx.is_charging);
     }
 
     #[test]
