+++ core/src/shared/mod.rs
@@ -1,0 +1,3 @@
+pub mod context;
+pub mod frontmatter;
+pub mod validated;
