+++ core/src/shared/validated.rs
